cohorts/2023/11-capstone-1/requirements.txt
fastapi
pandas
mlflow
hyperopt
python-decouple
prefect-azure[blob_storage]
uvicorn
gunicorn
